test: adjust tests
