+++ libs/bite-tribe/api/src/lib/restaurant-api/utils/get-restaurant-by-id.ts
